@@ -1,4 +1,4 @@
-class CreateUsersTable < ActiveRecord::Migration
+class CreateUsersTable < ActiveRecord::Migration[4.2]
   def self.up
     create_table :users do |t|
       t.string     :first_name,       :limit => 40
