@@ -1,4 +1,4 @@
-class AddContentTypeToJoinTables < ActiveRecord::Migration
+class AddContentTypeToJoinTables < ActiveRecord::Migration[4.2]
   def self.up
     add_column :category_assignments, :content_type, :string
     execute 'UPDATE category_assignments SET content_type = (SELECT type from contents WHERE contents.id = category_assignments.content_id)'
