@@ -1,4 +1,4 @@
-class AddPublishedAtToSections < ActiveRecord::Migration
+class AddPublishedAtToSections < ActiveRecord::Migration[4.2]
   def self.up
     add_column :sections, :published_at, :datetime
   end
