@@ -1,4 +1,4 @@
-class AddYearToRegistrations < ActiveRecord::Migration
+class AddYearToRegistrations < ActiveRecord::Migration[4.2]
   def self.up
     add_column :registrations, :year, :integer
   end
