@@ -1,4 +1,4 @@
-class AddPhoneToRiders < ActiveRecord::Migration
+class AddPhoneToRiders < ActiveRecord::Migration[4.2]
   def self.up
     add_column :riders, :phone, :string
   end
