@@ -1,4 +1,4 @@
-class AddSiteIdToTeams < ActiveRecord::Migration
+class AddSiteIdToTeams < ActiveRecord::Migration[4.2]
   def self.up
     add_column :teams, :site_id, :integer
   end
