@@ -1,4 +1,4 @@
-class AddRaceIdToTeams < ActiveRecord::Migration
+class AddRaceIdToTeams < ActiveRecord::Migration[4.2]
   def self.up
     add_column :teams, :race_id, :integer
   end
