@@ -1,4 +1,4 @@
-class AddStartTimeToRaces < ActiveRecord::Migration
+class AddStartTimeToRaces < ActiveRecord::Migration[4.2]
   def self.up
     add_column :races, :start_time, :datetime
   end
