@@ -1,4 +1,4 @@
-class AddRaceIdToPoints < ActiveRecord::Migration
+class AddRaceIdToPoints < ActiveRecord::Migration[4.2]
   def self.up
     add_column :points, :race_id, :integer
     add_index :points, :race_id
