@@ -1,4 +1,4 @@
-class AddDescriptionToRaces < ActiveRecord::Migration
+class AddDescriptionToRaces < ActiveRecord::Migration[4.2]
   def self.up
     add_column :races, :description, :text
   end
