@@ -1,4 +1,4 @@
-class AddPublishedToRaces < ActiveRecord::Migration
+class AddPublishedToRaces < ActiveRecord::Migration[4.2]
   def self.up
     add_column :races, :published, :boolean
   end
