@@ -1,4 +1,4 @@
-class AddBonusIdToPoints < ActiveRecord::Migration
+class AddBonusIdToPoints < ActiveRecord::Migration[4.2]
   def change
     add_column :points, :bonus_id, :integer
   end
